بروزرسانی پنل

+++ V2boardApi/Areas/App/Data/RequestModels/PlanResponseViewModel.cs
@@ -10,7 +10,7 @@ namespace V2boardApi.Areas.App.Data.RequestModels
         public int id { get; set; }
         public string PlanName { get; set; }
         public int DayesCount { get; set; }
-        public int Traffic { get; set; }
+        public double Traffic { get; set; }
         public string Price { get; set; }
         public int Status { get; set; }
         public string SpeedLimit { get; set; }
